tests/wpt/tests/html/rendering/widgets/select-wrap-no-spill.optional.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<table>
<tr><td id="target">
<select style="width:80px; white-space:pre-wrap;">
<option>ab option with a very long text that does not fit but should not spill</option>
</select>
<tr><td id="reference"><select style="width:80px;">
<option>ab option with a very long text that does not fit but should not spill</option>
</select>
</table>
<script>
// crbug.com/924929
test(() => {
  assert_equals(document.querySelector('#target').offsetHeight,
      document.querySelector('#reference').offsetHeight);
}, 'Selected OPTION label with white-space:pre-wrap should not spill out.');
</script>
</body>
</html>
